requirements.txt
pygame>=2.5.0
requests>=2.28.0
Pillow>=9.0.0
